refactor: EditableSelect 컴포넌트 common으로 옮기기
