+++ components/members/upload/AdditionalInfoFormSection.tsx
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 import { useFieldArray, useFormContext, useWatch } from 'react-hook-form';
 
+import EditableSelect from '@/components/common/EditableSelect';
 import Input from '@/components/common/Input';
 import Switch from '@/components/common/Switch';
 import Text from '@/components/common/Text';
@@ -10,7 +11,6 @@ import AddableWrapper from '@/components/members/upload/AddableWrapper';
 import { LINK_TITLES } from '@/components/members/upload/constants';
 import CountableInput from '@/components/members/upload/forms/CountableInput';
 import CountableTextArea from '@/components/members/upload/forms/CountableTextArea';
-import EditableSelect from '@/components/members/upload/forms/EditableSelect';
 import FormHeader from '@/components/members/upload/forms/FormHeader';
 import FormItem from '@/components/members/upload/forms/FormItem';
 import { MemberFormSection as FormSection } from '@/components/members/upload/forms/FormSection';
